Move node patch method to schema parser patch_node_properties and refactor schema parsing
