+++ core/dbt/parser/schema_yaml_readers.py
@@ -1,0 +1,271 @@
+from dbt.parser.schemas import YamlReader, SchemaParser
+from dbt.parser.common import YamlBlock
+from dbt.node_types import NodeType
+from dbt.contracts.graph.unparsed import UnparsedExposure, UnparsedMetric, UnparsedGroup
+from dbt.contracts.graph.nodes import Exposure, Metric, Group
+from dbt.exceptions import DbtInternalError, YamlParseDictError, JSONValidationError
+from dbt.context.providers import generate_parse_exposure, generate_parse_metrics
+from dbt.contracts.graph.model_config import MetricConfig, ExposureConfig
+from dbt.context.context_config import (
+    BaseContextConfigGenerator,
+    ContextConfigGenerator,
+    UnrenderedConfigGenerator,
+)
+from dbt.clients.jinja import get_rendered
+from typing import List
+from dbt.dataclass_schema import ValidationError
+
+
+class ExposureParser(YamlReader):
+    def __init__(self, schema_parser: SchemaParser, yaml: YamlBlock):
+        super().__init__(schema_parser, yaml, NodeType.Exposure.pluralize())
+        self.schema_parser = schema_parser
+        self.yaml = yaml
+
+    def parse_exposure(self, unparsed: UnparsedExposure):
+        package_name = self.project.project_name
+        unique_id = f"{NodeType.Exposure}.{package_name}.{unparsed.name}"
+        path = self.yaml.path.relative_path
+
+        fqn = self.schema_parser.get_fqn_prefix(path)
+        fqn.append(unparsed.name)
+
+        config = self._generate_exposure_config(
+            target=unparsed,
+            fqn=fqn,
+            package_name=package_name,
+            rendered=True,
+        )
+
+        config = config.finalize_and_validate()
+
+        unrendered_config = self._generate_exposure_config(
+            target=unparsed,
+            fqn=fqn,
+            package_name=package_name,
+            rendered=False,
+        )
+
+        if not isinstance(config, ExposureConfig):
+            raise DbtInternalError(
+                f"Calculated a {type(config)} for an exposure, but expected an ExposureConfig"
+            )
+
+        parsed = Exposure(
+            resource_type=NodeType.Exposure,
+            package_name=package_name,
+            path=path,
+            original_file_path=self.yaml.path.original_file_path,
+            unique_id=unique_id,
+            fqn=fqn,
+            name=unparsed.name,
+            type=unparsed.type,
+            url=unparsed.url,
+            meta=unparsed.meta,
+            tags=unparsed.tags,
+            description=unparsed.description,
+            label=unparsed.label,
+            owner=unparsed.owner,
+            maturity=unparsed.maturity,
+            config=config,
+            unrendered_config=unrendered_config,
+        )
+        ctx = generate_parse_exposure(
+            parsed,
+            self.root_project,
+            self.schema_parser.manifest,
+            package_name,
+        )
+        depends_on_jinja = "\n".join("{{ " + line + "}}" for line in unparsed.depends_on)
+        get_rendered(depends_on_jinja, ctx, parsed, capture_macros=True)
+        # parsed now has a populated refs/sources/metrics
+
+        if parsed.config.enabled:
+            self.manifest.add_exposure(self.yaml.file, parsed)
+        else:
+            self.manifest.add_disabled(self.yaml.file, parsed)
+
+    def _generate_exposure_config(
+        self, target: UnparsedExposure, fqn: List[str], package_name: str, rendered: bool
+    ):
+        generator: BaseContextConfigGenerator
+        if rendered:
+            generator = ContextConfigGenerator(self.root_project)
+        else:
+            generator = UnrenderedConfigGenerator(self.root_project)
+
+        # configs with precendence set
+        precedence_configs = dict()
+        # apply exposure configs
+        precedence_configs.update(target.config)
+
+        return generator.calculate_node_config(
+            config_call_dict={},
+            fqn=fqn,
+            resource_type=NodeType.Exposure,
+            project_name=package_name,
+            base=False,
+            patch_config_dict=precedence_configs,
+        )
+
+    def parse(self):
+        for data in self.get_key_dicts():
+            try:
+                UnparsedExposure.validate(data)
+                unparsed = UnparsedExposure.from_dict(data)
+            except (ValidationError, JSONValidationError) as exc:
+                raise YamlParseDictError(self.yaml.path, self.key, data, exc)
+
+            self.parse_exposure(unparsed)
+
+
+class MetricParser(YamlReader):
+    def __init__(self, schema_parser: SchemaParser, yaml: YamlBlock):
+        super().__init__(schema_parser, yaml, NodeType.Metric.pluralize())
+        self.schema_parser = schema_parser
+        self.yaml = yaml
+
+    def parse_metric(self, unparsed: UnparsedMetric):
+        package_name = self.project.project_name
+        unique_id = f"{NodeType.Metric}.{package_name}.{unparsed.name}"
+        path = self.yaml.path.relative_path
+
+        fqn = self.schema_parser.get_fqn_prefix(path)
+        fqn.append(unparsed.name)
+
+        config = self._generate_metric_config(
+            target=unparsed,
+            fqn=fqn,
+            package_name=package_name,
+            rendered=True,
+        )
+
+        config = config.finalize_and_validate()
+
+        unrendered_config = self._generate_metric_config(
+            target=unparsed,
+            fqn=fqn,
+            package_name=package_name,
+            rendered=False,
+        )
+
+        if not isinstance(config, MetricConfig):
+            raise DbtInternalError(
+                f"Calculated a {type(config)} for a metric, but expected a MetricConfig"
+            )
+
+        parsed = Metric(
+            resource_type=NodeType.Metric,
+            package_name=package_name,
+            path=path,
+            original_file_path=self.yaml.path.original_file_path,
+            unique_id=unique_id,
+            fqn=fqn,
+            model=unparsed.model,
+            name=unparsed.name,
+            description=unparsed.description,
+            label=unparsed.label,
+            calculation_method=unparsed.calculation_method,
+            expression=str(unparsed.expression),
+            timestamp=unparsed.timestamp,
+            dimensions=unparsed.dimensions,
+            window=unparsed.window,
+            time_grains=unparsed.time_grains,
+            filters=unparsed.filters,
+            meta=unparsed.meta,
+            tags=unparsed.tags,
+            config=config,
+            unrendered_config=unrendered_config,
+            group=config.group,
+        )
+
+        ctx = generate_parse_metrics(
+            parsed,
+            self.root_project,
+            self.schema_parser.manifest,
+            package_name,
+        )
+
+        if parsed.model is not None:
+            model_ref = "{{ " + parsed.model + " }}"
+            get_rendered(model_ref, ctx, parsed)
+
+        parsed.expression = get_rendered(
+            parsed.expression,
+            ctx,
+            node=parsed,
+        )
+
+        # if the metric is disabled we do not want it included in the manifest, only in the disabled dict
+        if parsed.config.enabled:
+            self.manifest.add_metric(self.yaml.file, parsed)
+        else:
+            self.manifest.add_disabled(self.yaml.file, parsed)
+
+    def _generate_metric_config(
+        self, target: UnparsedMetric, fqn: List[str], package_name: str, rendered: bool
+    ):
+        generator: BaseContextConfigGenerator
+        if rendered:
+            generator = ContextConfigGenerator(self.root_project)
+        else:
+            generator = UnrenderedConfigGenerator(self.root_project)
+
+        # configs with precendence set
+        precedence_configs = dict()
+        # first apply metric configs
+        precedence_configs.update(target.config)
+
+        config = generator.calculate_node_config(
+            config_call_dict={},
+            fqn=fqn,
+            resource_type=NodeType.Metric,
+            project_name=package_name,
+            base=False,
+            patch_config_dict=precedence_configs,
+        )
+        return config
+
+    def parse(self):
+        for data in self.get_key_dicts():
+            try:
+                UnparsedMetric.validate(data)
+                unparsed = UnparsedMetric.from_dict(data)
+
+            except (ValidationError, JSONValidationError) as exc:
+                raise YamlParseDictError(self.yaml.path, self.key, data, exc)
+            self.parse_metric(unparsed)
+
+
+class GroupParser(YamlReader):
+    def __init__(self, schema_parser: SchemaParser, yaml: YamlBlock):
+        super().__init__(schema_parser, yaml, NodeType.Group.pluralize())
+        self.schema_parser = schema_parser
+        self.yaml = yaml
+
+    def parse_group(self, unparsed: UnparsedGroup):
+        package_name = self.project.project_name
+        unique_id = f"{NodeType.Group}.{package_name}.{unparsed.name}"
+        path = self.yaml.path.relative_path
+
+        parsed = Group(
+            resource_type=NodeType.Group,
+            package_name=package_name,
+            path=path,
+            original_file_path=self.yaml.path.original_file_path,
+            unique_id=unique_id,
+            name=unparsed.name,
+            owner=unparsed.owner,
+        )
+
+        self.manifest.add_group(self.yaml.file, parsed)
+
+    def parse(self):
+        for data in self.get_key_dicts():
+            try:
+                UnparsedGroup.validate(data)
+                unparsed = UnparsedGroup.from_dict(data)
+            except (ValidationError, JSONValidationError) as exc:
+                raise YamlParseDictError(self.yaml.path, self.key, data, exc)
+
+            self.parse_group(unparsed)
