+++ core/dbt/parser/schemas.py
@@ -1,45 +1,21 @@
-import itertools
-import os
-import pathlib
+import time
 
 from abc import ABCMeta, abstractmethod
-from typing import Iterable, Dict, Any, Union, List, Optional, Generic, TypeVar, Type, Callable
+from typing import Iterable, Dict, Any, List, Generic, TypeVar, Type, Callable
 from dataclasses import dataclass, field
 
 from dbt.dataclass_schema import ValidationError, dbtClassMixin
 
-from dbt.adapters.factory import get_adapter, get_adapter_package_names
-from dbt.clients.jinja import get_rendered, add_rendered_test_kwargs
 from dbt.clients.yaml_helper import load_yaml_text
-from dbt.context.providers import RefArgs
 from dbt.parser.schema_renderer import SchemaYamlRenderer
-from dbt.context.context_config import (
-    ContextConfig,
-    BaseContextConfigGenerator,
-    ContextConfigGenerator,
-    UnrenderedConfigGenerator,
-)
+from dbt.parser.schema_generic_tests import SchemaGenericTestParser
+from dbt.context.context_config import ContextConfig
 from dbt.context.configured import generate_schema_yml_context, SchemaYamlVars
-from dbt.context.providers import (
-    generate_parse_exposure,
-    generate_parse_metrics,
-    generate_test_context,
-)
-from dbt.context.macro_resolver import MacroResolver
-from dbt.contracts.files import FileHash, SchemaSourceFile
-from dbt.contracts.graph.model_config import MetricConfig, ExposureConfig
+from dbt.contracts.files import SchemaSourceFile
 from dbt.contracts.graph.nodes import (
     ParsedNodePatch,
-    ColumnInfo,
-    ColumnLevelConstraint,
-    GenericTestNode,
     ParsedMacroPatch,
     UnpatchedSourceDefinition,
-    Exposure,
-    Metric,
-    Group,
-    ManifestNode,
-    GraphMemberNode,
     ConstraintType,
     ModelNode,
     ModelLevelConstraint,
@@ -47,53 +23,46 @@ from dbt.contracts.graph.nodes import (
 from dbt.contracts.graph.unparsed import (
     HasColumnDocs,
     HasColumnTests,
-    HasColumnProps,
     SourcePatch,
     UnparsedAnalysisUpdate,
-    UnparsedColumn,
     UnparsedMacroUpdate,
     UnparsedNodeUpdate,
     UnparsedModelUpdate,
-    UnparsedExposure,
-    UnparsedMetric,
     UnparsedSourceDefinition,
-    UnparsedGroup,
-    NodeVersion,
 )
 from dbt.exceptions import (
-    CompilationError,
     DuplicateMacroPatchNameError,
     DuplicatePatchPathError,
     DuplicateSourcePatchNameError,
     JSONValidationError,
     DbtInternalError,
-    SchemaConfigError,
-    TestConfigError,
     ParsingError,
     DbtValidationError,
     YamlLoadError,
     YamlParseDictError,
     YamlParseListError,
+    InvalidAccessTypeError,
 )
 from dbt.events.functions import warn_or_error
-from dbt.events.types import WrongResourceSchemaFile, NoNodeForYamlKey, MacroNotFoundForPatch
-from dbt.node_types import NodeType
+from dbt.events.types import (
+    WrongResourceSchemaFile,
+    NoNodeForYamlKey,
+    MacroNotFoundForPatch,
+    ValidationWarning,
+)
+from dbt.node_types import NodeType, AccessType
 from dbt.parser.base import SimpleParser
 from dbt.parser.search import FileBlock
-from dbt.parser.generic_test_builders import (
-    TestBuilder,
-    GenericTestBlock,
-    TargetBlock,
+from dbt.parser.common import (
     YamlBlock,
+    TargetBlock,
     TestBlock,
     VersionedTestBlock,
-    Testable,
-    Versioned,
+    ParserRef,
+    _trimmed,
 )
-from dbt.utils import get_pseudo_test_path, coerce_dict_str, md5, deep_merge
+from dbt.utils import coerce_dict_str, deep_merge
 
-
-TestDef = Union[str, Dict[str, Any]]
 
 schema_file_keys = (
     "models",
@@ -107,6 +76,36 @@ schema_file_keys = (
 )
 
 
+# ===============================================================================
+#  Schema Parser classes
+#
+# The SchemaParser is a subclass of the SimpleParser from base.py, as is
+# the SchemaGenericTestParser. The schema sub-parsers are all subclasses of
+# the YamlReader parsing class. Most of the action in creating SourceDefinition
+# nodes actually happens in the SourcePatcher class, in sources.py, which is
+# called as a late-stage parsing step in manifest.py.
+#
+# The "patch" parsers read yaml config and properties and apply them to
+# nodes that were already created from sql files.
+#
+# The SchemaParser and SourcePatcher both use the SchemaGenericTestParser
+# (in schema_generic_tests.py) to create generic test nodes.
+#
+#  YamlReader
+#      MetricParser (metrics) [schema_yaml_readers.py]
+#      ExposureParser (exposures) [schema_yaml_readers.py]
+#      GroupParser  (groups) [schema_yaml_readers.py]
+#      SourceParser (sources)
+#      PatchParser
+#          MacroPatchParser (macros)
+#          NodePatchParser
+#              ModelPatchParser (models)
+#              AnalysisPatchParser (analyses)
+#              TestablePatchParser (seeds, snapshots)
+#
+# ===============================================================================
+
+
 def yaml_from_file(source_file: SchemaSourceFile) -> Dict[str, Any]:
     """If loading the yaml fails, raise an exception."""
     try:
@@ -118,61 +117,9 @@ def yaml_from_file(source_file: SchemaSourceFile) -> Dict[str, Any]:
         )
 
 
-class ParserRef:
-    """A helper object to hold parse-time references."""
-
-    def __init__(self):
-        self.column_info: Dict[str, ColumnInfo] = {}
-
-    def _add(self, column: HasColumnProps):
-        tags: List[str] = []
-        tags.extend(getattr(column, "tags", ()))
-        quote: Optional[bool]
-        if isinstance(column, UnparsedColumn):
-            quote = column.quote
-        else:
-            quote = None
-
-        if any(
-            c
-            for c in column.constraints
-            if "type" not in c or not ConstraintType.is_valid(c["type"])
-        ):
-            raise ParsingError(f"Invalid constraint type on column {column.name}")
-
-        self.column_info[column.name] = ColumnInfo(
-            name=column.name,
-            description=column.description,
-            data_type=column.data_type,
-            constraints=[ColumnLevelConstraint.from_dict(c) for c in column.constraints],
-            meta=column.meta,
-            tags=tags,
-            quote=quote,
-            _extra=column.extra,
-        )
-
-    @classmethod
-    def from_target(cls, target: Union[HasColumnDocs, HasColumnTests]) -> "ParserRef":
-        refs = cls()
-        for column in target.columns:
-            refs._add(column)
-        return refs
-
-    @classmethod
-    def from_versioned_target(cls, target: Versioned, version: NodeVersion) -> "ParserRef":
-        refs = cls()
-        for base_column in target.get_columns_for_version(version):
-            refs._add(base_column)
-        return refs
-
-
-def _trimmed(inp: str) -> str:
-    if len(inp) < 50:
-        return inp
-    return inp[:44] + "..." + inp[-3:]
-
-
-class SchemaParser(SimpleParser[GenericTestBlock, GenericTestNode]):
+# This is the main schema file parser, but almost everything happens in the
+# the schema sub-parsers.
+class SchemaParser(SimpleParser[YamlBlock, ModelNode]):
     def __init__(
         self,
         project,
@@ -181,13 +128,11 @@ class SchemaParser(SimpleParser[GenericTestBlock, GenericTestNode]):
     ) -> None:
         super().__init__(project, manifest, root_project)
 
+        self.generic_test_parser = SchemaGenericTestParser(project, manifest, root_project)
+
         self.schema_yaml_vars = SchemaYamlVars()
         self.render_ctx = generate_schema_yml_context(
             self.root_project, self.project.project_name, self.schema_yaml_vars
-        )
-        internal_package_names = get_adapter_package_names(self.root_project.credentials.type)
-        self.macro_resolver = MacroResolver(
-            self.manifest.macros, self.root_project.project_name, internal_package_names
         )
 
     @classmethod
@@ -199,342 +144,6 @@ class SchemaParser(SimpleParser[GenericTestBlock, GenericTestNode]):
     def resource_type(self) -> NodeType:
         return NodeType.Test
 
-    def parse_from_dict(self, dct, validate=True) -> GenericTestNode:
-        if validate:
-            GenericTestNode.validate(dct)
-        return GenericTestNode.from_dict(dct)
-
-    def parse_column_tests(
-        self, block: TestBlock, column: UnparsedColumn, version: Optional[NodeVersion]
-    ) -> None:
-        if not column.tests:
-            return
-
-        for test in column.tests:
-            self.parse_test(block, test, column, version)
-
-    def create_test_node(
-        self,
-        target: Union[UnpatchedSourceDefinition, UnparsedNodeUpdate],
-        path: str,
-        config: ContextConfig,
-        tags: List[str],
-        fqn: List[str],
-        name: str,
-        raw_code: str,
-        test_metadata: Dict[str, Any],
-        file_key_name: str,
-        column_name: Optional[str],
-    ) -> GenericTestNode:
-
-        HASH_LENGTH = 10
-
-        # N.B: This function builds a hashable string from any given test_metadata dict.
-        #   it's a bit fragile for general use (only supports str, int, float, List, Dict)
-        #   but it gets the job done here without the overhead of complete ser(de).
-        def get_hashable_md(data: Union[str, int, float, List, Dict]) -> Union[str, List, Dict]:
-            if type(data) == dict:
-                return {k: get_hashable_md(data[k]) for k in sorted(data.keys())}  # type: ignore
-            elif type(data) == list:
-                return [get_hashable_md(val) for val in data]  # type: ignore
-            else:
-                return str(data)
-
-        hashable_metadata = repr(get_hashable_md(test_metadata))
-        hash_string = "".join([name, hashable_metadata])
-        test_hash = md5(hash_string)[-HASH_LENGTH:]
-
-        dct = {
-            "alias": name,
-            "schema": self.default_schema,
-            "database": self.default_database,
-            "fqn": fqn,
-            "name": name,
-            "resource_type": self.resource_type,
-            "tags": tags,
-            "path": path,
-            "original_file_path": target.original_file_path,
-            "package_name": self.project.project_name,
-            "raw_code": raw_code,
-            "language": "sql",
-            "unique_id": self.generate_unique_id(name, test_hash),
-            "config": self.config_dict(config),
-            "test_metadata": test_metadata,
-            "column_name": column_name,
-            "checksum": FileHash.empty().to_dict(omit_none=True),
-            "file_key_name": file_key_name,
-        }
-        try:
-            GenericTestNode.validate(dct)
-            return GenericTestNode.from_dict(dct)
-        except ValidationError as exc:
-            # this is a bit silly, but build an UnparsedNode just for error
-            # message reasons
-            node = self._create_error_node(
-                name=target.name,
-                path=path,
-                original_file_path=target.original_file_path,
-                raw_code=raw_code,
-            )
-            raise TestConfigError(exc, node)
-
-    # lots of time spent in this method
-    def _parse_generic_test(
-        self,
-        target: Testable,
-        test: Dict[str, Any],
-        tags: List[str],
-        column_name: Optional[str],
-        schema_file_id: str,
-        version: Optional[NodeVersion],
-    ) -> GenericTestNode:
-        try:
-            builder = TestBuilder(
-                test=test,
-                target=target,
-                column_name=column_name,
-                version=version,
-                package_name=target.package_name,
-                render_ctx=self.render_ctx,
-            )
-            if self.schema_yaml_vars.env_vars:
-                self.store_env_vars(target, schema_file_id, self.schema_yaml_vars.env_vars)
-                self.schema_yaml_vars.env_vars = {}
-
-        except ParsingError as exc:
-            context = _trimmed(str(target))
-            msg = "Invalid test config given in {}:\n\t{}\n\t@: {}".format(
-                target.original_file_path, exc.msg, context
-            )
-            raise ParsingError(msg) from exc
-
-        except CompilationError as exc:
-            context = _trimmed(str(target))
-            msg = (
-                "Invalid generic test configuration given in "
-                f"{target.original_file_path}: \n{exc.msg}\n\t@: {context}"
-            )
-            raise CompilationError(msg) from exc
-
-        original_name = os.path.basename(target.original_file_path)
-        compiled_path = get_pseudo_test_path(builder.compiled_name, original_name)
-
-        # fqn is the relative path of the yaml file where this generic test is defined,
-        # minus the project-level directory and the file name itself
-        # TODO pass a consistent path object from both UnparsedNode and UnpatchedSourceDefinition
-        path = pathlib.Path(target.original_file_path)
-        relative_path = str(path.relative_to(*path.parts[:1]))
-        fqn = self.get_fqn(relative_path, builder.fqn_name)
-
-        # this is the ContextConfig that is used in render_update
-        config: ContextConfig = self.initial_config(fqn)
-
-        # builder.args contains keyword args for the test macro,
-        # not configs which have been separated out in the builder.
-        # The keyword args are not completely rendered until compilation.
-        metadata = {
-            "namespace": builder.namespace,
-            "name": builder.name,
-            "kwargs": builder.args,
-        }
-        tags = sorted(set(itertools.chain(tags, builder.tags())))
-
-        if isinstance(target, UnpatchedSourceDefinition):
-            file_key_name = f"{target.source.yaml_key}.{target.source.name}"
-        else:
-            file_key_name = f"{target.yaml_key}.{target.name}"
-
-        node = self.create_test_node(
-            target=target,
-            path=compiled_path,
-            config=config,
-            fqn=fqn,
-            tags=tags,
-            name=builder.fqn_name,
-            raw_code=builder.build_raw_code(),
-            column_name=column_name,
-            test_metadata=metadata,
-            file_key_name=file_key_name,
-        )
-        self.render_test_update(node, config, builder, schema_file_id)
-
-        return node
-
-    def _lookup_attached_node(
-        self, target: Testable, version: Optional[NodeVersion]
-    ) -> Optional[Union[ManifestNode, GraphMemberNode]]:
-        """Look up attached node for Testable target nodes other than sources. Can be None if generic test attached to SQL node with no corresponding .sql file."""
-        attached_node = None  # type: Optional[Union[ManifestNode, GraphMemberNode]]
-        if not isinstance(target, UnpatchedSourceDefinition):
-            attached_node_unique_id = self.manifest.ref_lookup.get_unique_id(
-                target.name, None, version
-            )
-            if attached_node_unique_id:
-                attached_node = self.manifest.nodes[attached_node_unique_id]
-            else:
-                disabled_node = self.manifest.disabled_lookup.find(
-                    target.name, None
-                ) or self.manifest.disabled_lookup.find(target.name.upper(), None)
-                if disabled_node:
-                    attached_node = self.manifest.disabled[disabled_node[0].unique_id][0]
-        return attached_node
-
-    def store_env_vars(self, target, schema_file_id, env_vars):
-        self.manifest.env_vars.update(env_vars)
-        if schema_file_id in self.manifest.files:
-            schema_file = self.manifest.files[schema_file_id]
-            if isinstance(target, UnpatchedSourceDefinition):
-                search_name = target.source.name
-                yaml_key = target.source.yaml_key
-                if "." in search_name:  # source file definitions
-                    (search_name, _) = search_name.split(".")
-            else:
-                search_name = target.name
-                yaml_key = target.yaml_key
-            for var in env_vars.keys():
-                schema_file.add_env_var(var, yaml_key, search_name)
-
-    # This does special shortcut processing for the two
-    # most common internal macros, not_null and unique,
-    # which avoids the jinja rendering to resolve config
-    # and variables, etc, which might be in the macro.
-    # In the future we will look at generalizing this
-    # more to handle additional macros or to use static
-    # parsing to avoid jinja overhead.
-    def render_test_update(self, node, config, builder, schema_file_id):
-        macro_unique_id = self.macro_resolver.get_macro_id(
-            node.package_name, "test_" + builder.name
-        )
-        # Add the depends_on here so we can limit the macros added
-        # to the context in rendering processing
-        node.depends_on.add_macro(macro_unique_id)
-        if macro_unique_id in ["macro.dbt.test_not_null", "macro.dbt.test_unique"]:
-            config_call_dict = builder.get_static_config()
-            config._config_call_dict = config_call_dict
-            # This sets the config from dbt_project
-            self.update_parsed_node_config(node, config)
-            # source node tests are processed at patch_source time
-            if isinstance(builder.target, UnpatchedSourceDefinition):
-                sources = [builder.target.fqn[-2], builder.target.fqn[-1]]
-                node.sources.append(sources)
-            else:  # all other nodes
-                node.refs.append(RefArgs(name=builder.target.name, version=builder.version))
-        else:
-            try:
-                # make a base context that doesn't have the magic kwargs field
-                context = generate_test_context(
-                    node,
-                    self.root_project,
-                    self.manifest,
-                    config,
-                    self.macro_resolver,
-                )
-                # update with rendered test kwargs (which collects any refs)
-                # Note: This does not actually update the kwargs with the rendered
-                # values. That happens in compilation.
-                add_rendered_test_kwargs(context, node, capture_macros=True)
-                # the parsed node is not rendered in the native context.
-                get_rendered(node.raw_code, context, node, capture_macros=True)
-                self.update_parsed_node_config(node, config)
-                # env_vars should have been updated in the context env_var method
-            except ValidationError as exc:
-                # we got a ValidationError - probably bad types in config()
-                raise SchemaConfigError(exc, node=node) from exc
-
-        # Set attached_node for generic test nodes, if available.
-        # Generic test node inherits attached node's group config value.
-        attached_node = self._lookup_attached_node(builder.target, builder.version)
-        if attached_node:
-            node.attached_node = attached_node.unique_id
-            node.group, node.group = attached_node.group, attached_node.group
-
-    def parse_node(self, block: GenericTestBlock) -> GenericTestNode:
-        """In schema parsing, we rewrite most of the part of parse_node that
-        builds the initial node to be parsed, but rendering is basically the
-        same
-        """
-        node = self._parse_generic_test(
-            target=block.target,
-            test=block.test,
-            tags=block.tags,
-            column_name=block.column_name,
-            schema_file_id=block.file.file_id,
-            version=block.version,
-        )
-        self.add_test_node(block, node)
-        return node
-
-    def add_test_node(self, block: GenericTestBlock, node: GenericTestNode):
-        test_from = {"key": block.target.yaml_key, "name": block.target.name}
-        if node.config.enabled:
-            self.manifest.add_node(block.file, node, test_from)
-        else:
-            self.manifest.add_disabled(block.file, node, test_from)
-
-    def render_with_context(
-        self,
-        node: GenericTestNode,
-        config: ContextConfig,
-    ) -> None:
-        """Given the parsed node and a ContextConfig to use during
-        parsing, collect all the refs that might be squirreled away in the test
-        arguments. This includes the implicit "model" argument.
-        """
-        # make a base context that doesn't have the magic kwargs field
-        context = self._context_for(node, config)
-        # update it with the rendered test kwargs (which collects any refs)
-        add_rendered_test_kwargs(context, node, capture_macros=True)
-
-        # the parsed node is not rendered in the native context.
-        get_rendered(node.raw_code, context, node, capture_macros=True)
-
-    def parse_test(
-        self,
-        target_block: TestBlock,
-        test: TestDef,
-        column: Optional[UnparsedColumn],
-        version: Optional[NodeVersion],
-    ) -> None:
-        if isinstance(test, str):
-            test = {test: {}}
-
-        if column is None:
-            column_name: Optional[str] = None
-            column_tags: List[str] = []
-        else:
-            column_name = column.name
-            should_quote = column.quote or (column.quote is None and target_block.quote_columns)
-            if should_quote:
-                column_name = get_adapter(self.root_project).quote(column_name)
-            column_tags = column.tags
-
-        block = GenericTestBlock.from_test_block(
-            src=target_block,
-            test=test,
-            column_name=column_name,
-            tags=column_tags,
-            version=version,
-        )
-        self.parse_node(block)
-
-    def parse_tests(self, block: TestBlock) -> None:
-        for column in block.columns:
-            self.parse_column_tests(block, column, None)
-
-        for test in block.tests:
-            self.parse_test(block, test, None, None)
-
-    def parse_versioned_tests(self, block: VersionedTestBlock) -> None:
-        if not block.target.versions:
-            self.parse_tests(block)
-        else:
-            for version in block.target.versions:
-                for column in block.target.get_columns_for_version(version.v):
-                    self.parse_column_tests(block, column, version.v)
-
-                for test in block.target.get_tests_for_version(version.v):
-                    self.parse_test(block, test, None, version.v)
-
     def parse_file(self, block: FileBlock, dct: Dict = None) -> None:
         assert isinstance(block.file, SchemaSourceFile)
 
@@ -544,10 +153,10 @@ class SchemaParser(SimpleParser[GenericTestBlock, GenericTestNode]):
             # contains the FileBlock and the data (dictionary)
             yaml_block = YamlBlock.from_file_block(block, dct)
 
-            parser: YamlDocsReader
+            parser: YamlReader
 
-            # There are 7 kinds of parsers:
-            # Model, Seed, Snapshot, Source, Macro, Analysis, Exposures
+            # There are 9 different yaml lists which are parsed by different parsers:
+            # Model, Seed, Snapshot, Source, Macro, Analysis, Exposure, Metric, Group
 
             # ModelPatchParser.parse()
             if "models" in dct:
@@ -555,19 +164,19 @@ class SchemaParser(SimpleParser[GenericTestBlock, GenericTestNode]):
                 # even if they are disabled in the schema file
                 model_parse_result = ModelPatchParser(self, yaml_block, "models").parse()
                 for versioned_test_block in model_parse_result.versioned_test_blocks:
-                    self.parse_versioned_tests(versioned_test_block)
+                    self.generic_test_parser.parse_versioned_tests(versioned_test_block)
 
-            # NonSourceParser.parse()
+            # PatchParser.parse()
             if "seeds" in dct:
                 seed_parse_result = TestablePatchParser(self, yaml_block, "seeds").parse()
                 for test_block in seed_parse_result.test_blocks:
-                    self.parse_tests(test_block)
+                    self.generic_test_parser.parse_tests(test_block)
 
-            # NonSourceParser.parse()
+            # PatchParser.parse()
             if "snapshots" in dct:
                 snapshot_parse_result = TestablePatchParser(self, yaml_block, "snapshots").parse()
                 for test_block in snapshot_parse_result.test_blocks:
-                    self.parse_tests(test_block)
+                    self.generic_test_parser.parse_tests(test_block)
 
             # This parser uses SourceParser.parse() which doesn't return
             # any test blocks. Source tests are handled at a later point
@@ -576,28 +185,34 @@ class SchemaParser(SimpleParser[GenericTestBlock, GenericTestNode]):
                 parser = SourceParser(self, yaml_block, "sources")
                 parser.parse()
 
-            # NonSourceParser.parse() (but never test_blocks)
+            # PatchParser.parse() (but never test_blocks)
             if "macros" in dct:
                 parser = MacroPatchParser(self, yaml_block, "macros")
                 parser.parse()
 
-            # NonSourceParser.parse() (but never test_blocks)
+            # PatchParser.parse() (but never test_blocks)
             if "analyses" in dct:
                 parser = AnalysisPatchParser(self, yaml_block, "analyses")
                 parser.parse()
 
-            # parse exposures
+            # ExposureParser.parse()
             if "exposures" in dct:
+                from dbt.parser.schema_yaml_readers import ExposureParser
+
                 exp_parser = ExposureParser(self, yaml_block)
                 exp_parser.parse()
 
-            # parse metrics
+            # MetricParser.parse()
             if "metrics" in dct:
+                from dbt.parser.schema_yaml_readers import MetricParser
+
                 metric_parser = MetricParser(self, yaml_block)
                 metric_parser.parse()
 
-            # parse groups
+            # GroupParser.parse()
             if "groups" in dct:
+                from dbt.parser.schema_yaml_readers import GroupParser
+
                 group_parser = GroupParser(self, yaml_block)
                 group_parser.parse()
 
@@ -620,6 +235,7 @@ class ParseResult:
 
 
 # abstract base class (ABCMeta)
+# Four subclasses: MetricParser, ExposureParser, GroupParser, SourceParser, PatchParser
 class YamlReader(metaclass=ABCMeta):
     def __init__(self, schema_parser: SchemaParser, yaml: YamlBlock, key: str) -> None:
         self.schema_parser = schema_parser
@@ -699,8 +315,6 @@ class YamlReader(metaclass=ABCMeta):
             ) from exc
         return dct
 
-
-class YamlDocsReader(YamlReader):
     @abstractmethod
     def parse(self) -> ParseResult:
         raise NotImplementedError("parse is abstract")
@@ -710,7 +324,7 @@ T = TypeVar("T", bound=dbtClassMixin)
 
 
 # This parses the 'sources' keys in yaml files.
-class SourceParser(YamlDocsReader):
+class SourceParser(YamlReader):
     def _target_from_dict(self, cls: Type[T], data: Dict[str, Any]) -> T:
         path = self.yaml.path.original_file_path
         try:
@@ -719,8 +333,7 @@ class SourceParser(YamlDocsReader):
         except (ValidationError, JSONValidationError) as exc:
             raise YamlParseDictError(path, self.key, data, exc)
 
-    # The other parse method returns TestBlocks. This one doesn't.
-    # This takes the yaml dictionaries in 'sources' keys and uses them
+    # This parse method takes the yaml dictionaries in 'sources' keys and uses them
     # to create UnparsedSourceDefinition objects. They are then turned
     # into UnpatchedSourceDefinition objects in 'add_source_definitions'
     # or SourcePatch objects in 'add_source_patch'
@@ -771,9 +384,8 @@ class SourceParser(YamlDocsReader):
             self.manifest.add_source(self.yaml.file, source_def)
 
 
-# This class has three main subclasses: TestablePatchParser (models,
-# seeds, snapshots), MacroPatchParser, and AnalysisPatchParser
-class NonSourceParser(YamlDocsReader, Generic[NonSourceTarget, Parsed]):
+# This class has two subclasses: NodePatchParser and MacroPatchParser
+class PatchParser(YamlReader, Generic[NonSourceTarget, Parsed]):
     @abstractmethod
     def _target_type(self) -> Type[NonSourceTarget]:
         raise NotImplementedError("_target_type not implemented")
@@ -896,7 +508,9 @@ class NonSourceParser(YamlDocsReader, Generic[NonSourceTarget, Parsed]):
         self.schema_parser.update_parsed_node_config(node, config, patch_config_dict=patch.config)
 
 
-class NodePatchParser(NonSourceParser[NodeTarget, ParsedNodePatch], Generic[NodeTarget]):
+# Subclasses of NodePatchParser: TestablePatchParser, ModelPatchParser, AnalysisPatchParser,
+# so models, seeds, snapshots, analyses
+class NodePatchParser(PatchParser[NodeTarget, ParsedNodePatch], Generic[NodeTarget]):
     def parse_patch(self, block: TargetBlock[NodeTarget], refs: ParserRef) -> None:
         # We're not passing the ParsedNodePatch around anymore, so we
         # could possibly skip creating one. Leaving here for now for
@@ -914,6 +528,7 @@ class NodePatchParser(NonSourceParser[NodeTarget, ParsedNodePatch], Generic[Node
             access=block.target.access,
             version=None,
             latest_version=None,
+            constraints=block.target.constraints,
         )
         assert isinstance(self.yaml.file, SchemaSourceFile)
         source_file: SchemaSourceFile = self.yaml.file
@@ -970,7 +585,7 @@ class NodePatchParser(NonSourceParser[NodeTarget, ParsedNodePatch], Generic[Node
                     if patch.config:
                         self.patch_node_config(node, patch)
 
-                    node.patch(patch)
+                    self.patch_node_properties(node, patch)
             else:
                 warn_or_error(
                     NoNodeForYamlKey(
@@ -994,48 +609,30 @@ class NodePatchParser(NonSourceParser[NodeTarget, ParsedNodePatch], Generic[Node
             if patch.config:
                 self.patch_node_config(node, patch)
 
-            node.patch(patch)
-            # TODO: We want to do all the actual patching either in the above node.patch() call
-            # or here, but it will require some thought to the details. For now the patching is
-            # awkwardly split.
-            self.patch_constraints(node, block.target.constraints)
-            node.build_contract_checksum()
+            self.patch_node_properties(node, patch)
 
-    def patch_constraints(self, node, constraints):
-        contract_config = node.config.get("contract")
-        if isinstance(node, ModelNode) and contract_config.enforced is True:
-            self._validate_constraint_prerequisites(node)
+    def patch_node_properties(self, node, patch: "ParsedNodePatch"):
+        """Given a ParsedNodePatch, add the new information to the node."""
+        # explicitly pick out the parts to update so we don't inadvertently
+        # step on the model name or anything
+        # Note: config should already be updated
+        node.patch_path = patch.file_id
+        # update created_at so process_docs will run in partial parsing
+        node.created_at = time.time()
+        node.description = patch.description
+        node.columns = patch.columns
+        node.name = patch.name
 
-            if any(
-                c for c in constraints if "type" not in c or not ConstraintType.is_valid(c["type"])
-            ):
-                raise ParsingError(
-                    f"Invalid constraint type on model {node.name}: "
-                    f"Type must be one of {[ct.value for ct in ConstraintType]}"
-                )
-
-            node.constraints = [ModelLevelConstraint.from_dict(c) for c in constraints]
-
-    def _validate_constraint_prerequisites(self, model_node: ModelNode):
-        errors = []
-        if not model_node.columns:
-            errors.append(
-                "Constraints must be defined in a `yml` schema configuration file like `schema.yml`."
-            )
-
-        if model_node.config.materialized not in ["table", "view", "incremental"]:
-            errors.append(
-                f"Only table, view, and incremental materializations are supported for constraints, but found '{model_node.config.materialized}'"
-            )
-
-        if str(model_node.language) != "sql":
-            errors.append(f"Language Error: Expected 'sql' but found '{model_node.language}'")
-
-        if errors:
-            raise ParsingError(
-                f"Constraint validation failed for: ({model_node.original_file_path})\n"
-                + "\n".join(errors)
-            )
+        if not isinstance(node, ModelNode):
+            for attr in ["latest_version", "access", "version", "constraints"]:
+                if getattr(patch, attr):
+                    warn_or_error(
+                        ValidationWarning(
+                            field_name=attr,
+                            resource_type=node.resource_type.value,
+                            node_name=patch.name,
+                        )
+                    )
 
 
 # TestablePatchParser = seeds, snapshots
@@ -1163,17 +760,12 @@ class ModelPatchParser(NodePatchParser[UnparsedModelUpdate]):
                     access=unparsed_version.access or target.access,
                     version=unparsed_version.v,
                     latest_version=latest_version,
+                    constraints=unparsed_version.constraints or target.constraints,
                 )
                 # Node patched before config because config patching depends on model name,
                 # which may have been updated in the version patch
-                versioned_model_node.patch(versioned_model_patch)
-                # TODO: We want to do all the actual patching either in the above node.patch() call
-                # or here, but it will require some thought to the details. For now the patching is
-                # awkwardly split.
-                self.patch_constraints(
-                    versioned_model_node, unparsed_version.constraints or target.constraints
-                )
-                versioned_model_node.build_contract_checksum()
+                # versioned_model_node.patch(versioned_model_patch)
+                self.patch_node_properties(versioned_model_node, versioned_model_patch)
 
                 # Includes alias recomputation
                 self.patch_node_config(versioned_model_node, versioned_model_patch)
@@ -1186,6 +778,57 @@ class ModelPatchParser(NodePatchParser[UnparsedModelUpdate]):
     def _target_type(self) -> Type[UnparsedModelUpdate]:
         return UnparsedModelUpdate
 
+    def patch_node_properties(self, node, patch: "ParsedNodePatch"):
+        super().patch_node_properties(node, patch)
+        node.version = patch.version
+        node.latest_version = patch.latest_version
+        if patch.access:
+            if AccessType.is_valid(patch.access):
+                node.access = AccessType(patch.access)
+            else:
+                raise InvalidAccessTypeError(
+                    unique_id=node.unique_id,
+                    field_value=patch.access,
+                )
+        self.patch_constraints(node, patch.constraints)
+        node.build_contract_checksum()
+
+    def patch_constraints(self, node, constraints):
+        contract_config = node.config.get("contract")
+        if contract_config.enforced is True:
+            self._validate_constraint_prerequisites(node)
+
+            if any(
+                c for c in constraints if "type" not in c or not ConstraintType.is_valid(c["type"])
+            ):
+                raise ParsingError(
+                    f"Invalid constraint type on model {node.name}: "
+                    f"Type must be one of {[ct.value for ct in ConstraintType]}"
+                )
+
+            node.constraints = [ModelLevelConstraint.from_dict(c) for c in constraints]
+
+    def _validate_constraint_prerequisites(self, model_node: ModelNode):
+        errors = []
+        if not model_node.columns:
+            errors.append(
+                "Constraints must be defined in a `yml` schema configuration file like `schema.yml`."
+            )
+
+        if model_node.config.materialized not in ["table", "view", "incremental"]:
+            errors.append(
+                f"Only table, view, and incremental materializations are supported for constraints, but found '{model_node.config.materialized}'"
+            )
+
+        if str(model_node.language) != "sql":
+            errors.append(f"Language Error: Expected 'sql' but found '{model_node.language}'")
+
+        if errors:
+            raise ParsingError(
+                f"Constraint validation failed for: ({model_node.original_file_path})\n"
+                + "\n".join(errors)
+            )
+
 
 class AnalysisPatchParser(NodePatchParser[UnparsedAnalysisUpdate]):
     def get_block(self, node: UnparsedAnalysisUpdate) -> TargetBlock:
@@ -1195,7 +838,7 @@ class AnalysisPatchParser(NodePatchParser[UnparsedAnalysisUpdate]):
         return UnparsedAnalysisUpdate
 
 
-class MacroPatchParser(NonSourceParser[UnparsedMacroUpdate, ParsedMacroPatch]):
+class MacroPatchParser(PatchParser[UnparsedMacroUpdate, ParsedMacroPatch]):
     def get_block(self, node: UnparsedMacroUpdate) -> TargetBlock:
         return TargetBlock.from_yaml_block(self.yaml, node)
 
@@ -1226,259 +869,11 @@ class MacroPatchParser(NonSourceParser[UnparsedMacroUpdate, ParsedMacroPatch]):
             package_name, existing_file_path = macro.patch_path.split("://")
             raise DuplicateMacroPatchNameError(patch, existing_file_path)
         source_file.macro_patches[patch.name] = unique_id
-        macro.patch(patch)
 
-
-class ExposureParser(YamlReader):
-    def __init__(self, schema_parser: SchemaParser, yaml: YamlBlock):
-        super().__init__(schema_parser, yaml, NodeType.Exposure.pluralize())
-        self.schema_parser = schema_parser
-        self.yaml = yaml
-
-    def parse_exposure(self, unparsed: UnparsedExposure):
-        package_name = self.project.project_name
-        unique_id = f"{NodeType.Exposure}.{package_name}.{unparsed.name}"
-        path = self.yaml.path.relative_path
-
-        fqn = self.schema_parser.get_fqn_prefix(path)
-        fqn.append(unparsed.name)
-
-        config = self._generate_exposure_config(
-            target=unparsed,
-            fqn=fqn,
-            package_name=package_name,
-            rendered=True,
-        )
-
-        config = config.finalize_and_validate()
-
-        unrendered_config = self._generate_exposure_config(
-            target=unparsed,
-            fqn=fqn,
-            package_name=package_name,
-            rendered=False,
-        )
-
-        if not isinstance(config, ExposureConfig):
-            raise DbtInternalError(
-                f"Calculated a {type(config)} for an exposure, but expected an ExposureConfig"
-            )
-
-        parsed = Exposure(
-            resource_type=NodeType.Exposure,
-            package_name=package_name,
-            path=path,
-            original_file_path=self.yaml.path.original_file_path,
-            unique_id=unique_id,
-            fqn=fqn,
-            name=unparsed.name,
-            type=unparsed.type,
-            url=unparsed.url,
-            meta=unparsed.meta,
-            tags=unparsed.tags,
-            description=unparsed.description,
-            label=unparsed.label,
-            owner=unparsed.owner,
-            maturity=unparsed.maturity,
-            config=config,
-            unrendered_config=unrendered_config,
-        )
-        ctx = generate_parse_exposure(
-            parsed,
-            self.root_project,
-            self.schema_parser.manifest,
-            package_name,
-        )
-        depends_on_jinja = "\n".join("{{ " + line + "}}" for line in unparsed.depends_on)
-        get_rendered(depends_on_jinja, ctx, parsed, capture_macros=True)
-        # parsed now has a populated refs/sources/metrics
-
-        if parsed.config.enabled:
-            self.manifest.add_exposure(self.yaml.file, parsed)
-        else:
-            self.manifest.add_disabled(self.yaml.file, parsed)
-
-    def _generate_exposure_config(
-        self, target: UnparsedExposure, fqn: List[str], package_name: str, rendered: bool
-    ):
-        generator: BaseContextConfigGenerator
-        if rendered:
-            generator = ContextConfigGenerator(self.root_project)
-        else:
-            generator = UnrenderedConfigGenerator(self.root_project)
-
-        # configs with precendence set
-        precedence_configs = dict()
-        # apply exposure configs
-        precedence_configs.update(target.config)
-
-        return generator.calculate_node_config(
-            config_call_dict={},
-            fqn=fqn,
-            resource_type=NodeType.Exposure,
-            project_name=package_name,
-            base=False,
-            patch_config_dict=precedence_configs,
-        )
-
-    def parse(self):
-        for data in self.get_key_dicts():
-            try:
-                UnparsedExposure.validate(data)
-                unparsed = UnparsedExposure.from_dict(data)
-            except (ValidationError, JSONValidationError) as exc:
-                raise YamlParseDictError(self.yaml.path, self.key, data, exc)
-
-            self.parse_exposure(unparsed)
-
-
-class MetricParser(YamlReader):
-    def __init__(self, schema_parser: SchemaParser, yaml: YamlBlock):
-        super().__init__(schema_parser, yaml, NodeType.Metric.pluralize())
-        self.schema_parser = schema_parser
-        self.yaml = yaml
-
-    def parse_metric(self, unparsed: UnparsedMetric):
-        package_name = self.project.project_name
-        unique_id = f"{NodeType.Metric}.{package_name}.{unparsed.name}"
-        path = self.yaml.path.relative_path
-
-        fqn = self.schema_parser.get_fqn_prefix(path)
-        fqn.append(unparsed.name)
-
-        config = self._generate_metric_config(
-            target=unparsed,
-            fqn=fqn,
-            package_name=package_name,
-            rendered=True,
-        )
-
-        config = config.finalize_and_validate()
-
-        unrendered_config = self._generate_metric_config(
-            target=unparsed,
-            fqn=fqn,
-            package_name=package_name,
-            rendered=False,
-        )
-
-        if not isinstance(config, MetricConfig):
-            raise DbtInternalError(
-                f"Calculated a {type(config)} for a metric, but expected a MetricConfig"
-            )
-
-        parsed = Metric(
-            resource_type=NodeType.Metric,
-            package_name=package_name,
-            path=path,
-            original_file_path=self.yaml.path.original_file_path,
-            unique_id=unique_id,
-            fqn=fqn,
-            model=unparsed.model,
-            name=unparsed.name,
-            description=unparsed.description,
-            label=unparsed.label,
-            calculation_method=unparsed.calculation_method,
-            expression=str(unparsed.expression),
-            timestamp=unparsed.timestamp,
-            dimensions=unparsed.dimensions,
-            window=unparsed.window,
-            time_grains=unparsed.time_grains,
-            filters=unparsed.filters,
-            meta=unparsed.meta,
-            tags=unparsed.tags,
-            config=config,
-            unrendered_config=unrendered_config,
-            group=config.group,
-        )
-
-        ctx = generate_parse_metrics(
-            parsed,
-            self.root_project,
-            self.schema_parser.manifest,
-            package_name,
-        )
-
-        if parsed.model is not None:
-            model_ref = "{{ " + parsed.model + " }}"
-            get_rendered(model_ref, ctx, parsed)
-
-        parsed.expression = get_rendered(
-            parsed.expression,
-            ctx,
-            node=parsed,
-        )
-
-        # if the metric is disabled we do not want it included in the manifest, only in the disabled dict
-        if parsed.config.enabled:
-            self.manifest.add_metric(self.yaml.file, parsed)
-        else:
-            self.manifest.add_disabled(self.yaml.file, parsed)
-
-    def _generate_metric_config(
-        self, target: UnparsedMetric, fqn: List[str], package_name: str, rendered: bool
-    ):
-        generator: BaseContextConfigGenerator
-        if rendered:
-            generator = ContextConfigGenerator(self.root_project)
-        else:
-            generator = UnrenderedConfigGenerator(self.root_project)
-
-        # configs with precendence set
-        precedence_configs = dict()
-        # first apply metric configs
-        precedence_configs.update(target.config)
-
-        config = generator.calculate_node_config(
-            config_call_dict={},
-            fqn=fqn,
-            resource_type=NodeType.Metric,
-            project_name=package_name,
-            base=False,
-            patch_config_dict=precedence_configs,
-        )
-        return config
-
-    def parse(self):
-        for data in self.get_key_dicts():
-            try:
-                UnparsedMetric.validate(data)
-                unparsed = UnparsedMetric.from_dict(data)
-
-            except (ValidationError, JSONValidationError) as exc:
-                raise YamlParseDictError(self.yaml.path, self.key, data, exc)
-            self.parse_metric(unparsed)
-
-
-class GroupParser(YamlReader):
-    def __init__(self, schema_parser: SchemaParser, yaml: YamlBlock):
-        super().__init__(schema_parser, yaml, NodeType.Group.pluralize())
-        self.schema_parser = schema_parser
-        self.yaml = yaml
-
-    def parse_group(self, unparsed: UnparsedGroup):
-        package_name = self.project.project_name
-        unique_id = f"{NodeType.Group}.{package_name}.{unparsed.name}"
-        path = self.yaml.path.relative_path
-
-        parsed = Group(
-            resource_type=NodeType.Group,
-            package_name=package_name,
-            path=path,
-            original_file_path=self.yaml.path.original_file_path,
-            unique_id=unique_id,
-            name=unparsed.name,
-            owner=unparsed.owner,
-        )
-
-        self.manifest.add_group(self.yaml.file, parsed)
-
-    def parse(self):
-        for data in self.get_key_dicts():
-            try:
-                UnparsedGroup.validate(data)
-                unparsed = UnparsedGroup.from_dict(data)
-            except (ValidationError, JSONValidationError) as exc:
-                raise YamlParseDictError(self.yaml.path, self.key, data, exc)
-
-            self.parse_group(unparsed)
+        # former macro.patch code
+        macro.patch_path = patch.file_id
+        macro.description = patch.description
+        macro.created_at = time.time()
+        macro.meta = patch.meta
+        macro.docs = patch.docs
+        macro.arguments = patch.arguments
